move application related modules to common modules folder
