+++ Src/dronecan_application/CMakeLists.txt
@@ -3,7 +3,8 @@
 
 set(APPLICATION_DIR ${CMAKE_CURRENT_LIST_DIR})
 cmake_path(GET CMAKE_CURRENT_LIST_DIR PARENT_PATH SRC_DIR)
+set(MODULES_DIR ${SRC_DIR}/modules)
 
-include(${SRC_DIR}/modules/dronecan_module/CMakeLists.txt)
-include(${APPLICATION_DIR}/modules/circuit_status/CMakeLists.txt)
-include(${APPLICATION_DIR}/modules/pwm/CMakeLists.txt)
+include(${MODULES_DIR}/dronecan/core/CMakeLists.txt)
+include(${MODULES_DIR}/dronecan/circuit_status/CMakeLists.txt)
+include(${MODULES_DIR}/dronecan/pwm/CMakeLists.txt)
