+++ Src/modules/cyphal/circuit_status/circuit_status.hpp
